--- Content.Client/Foldable/FoldableSystem.cs
@@ -1,10 +1,0 @@
-using Content.Shared.Foldable;
-using JetBrains.Annotations;
-
-namespace Content.Client.Foldable;
-
-[UsedImplicitly]
-public sealed class FoldableSystem : SharedFoldableSystem
-{
-    // classic.
-}
